requirements.txt
asgiref==3.10.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
dj-database-url==3.0.1
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
Flask==3.1.2
flask-cors==6.0.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.4.0
MarkupSafe==3.0.3
mysqlclient==2.2.7
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-engineio==4.12.3
python-socketio==5.13.0
pytz==2025.2
requests==2.32.5
simple-websocket==1.1.0
simplekml==1.3.6
six==1.17.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.11.0
wsproto==1.2.0
xlrd==2.0.2
